Change Artifact ID to Jenkins.
